Add files via upload
This code aims to generate ECG signals for the 12 leads: unipolar leads aVR, aVL, and aVF, bipolar leads DI, DII, and DIII, and precordial leads V1, V2, V3, V4, V5, V6. The generated ECG signals are artificial, created from sinusoidal functions of different orders. Once the signals are generated, they are plotted for the 12 leads in a format similar to a real electrocardiogram exam, and the signals are saved within a struct. The code is still under development and is currently a version designed only to generate signals with a normal sinus rhythm of 60 bpm.